requirements.txt
watchdog
python-Levenshtein
